backend/requirements.txt
pymongo==3.11.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
python-dotenv==0.19.0